+++ server/pages/structured_json.py
@@ -1,0 +1,7 @@
+import streamlit as st
+
+st.title("LeanAide: Structured Json Output")
+st.write("Here you can input your theorem-proof/paper, etc. and generate structured JSON output using LeanAide Schema's.")
+
+st.sidebar.header("Structured Json")
+st.write("TBD")
