+++ server/pyproject.toml
@@ -7,6 +7,7 @@ requires-python = ">=3.13"
 dependencies = [
     "pyperclip>=1.9.0",
     "python-dotenv>=1.1.0",
+    "streamlit-code-editor>=0.1.22",
     "streamlit>=1.45.1",
     "watchdog>=6.0.0",
 ]
